requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
mysql-connector-python==8.1.0
sqlalchemy==2.0.20
pymysql==1.1.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pypdf2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2
pillow==10.0.0
librosa==0.10.1
opencv-python==4.8.1.78
scikit-learn==1.3.0
scipy==1.11.3
transformers==4.33.3
torch==2.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
langchain==0.0.287
google-generativeai==0.3.0
plotly==5.15.0
seaborn==0.12.2
matplotlib==3.7.2
altair==5.0.1
great-expectations==0.17.6
prefect==2.10.5
apache-airflow==2.7.1
python-magic==0.4.27
nltk==3.8.1